chore: 메모 스타일 코드 분리
