+++ src/components/learnDetail/memo/Memo/index.tsx
@@ -1,9 +1,7 @@
-import styled from 'styled-components';
-import { COLOR } from '@src/styles/color';
-import { FONT_STYLES } from '@src/styles/fontStyle';
 import { useState } from 'react';
-import MemoForm from './MemoForm';
-import MemoDotButton from './MemoPopupButton';
+import MemoForm from '../MemoForm';
+import MemoDotButton from '../MemoPopupButton';
+import { StContent, StKeyword, StMemo } from './style';
 
 interface MemoProps {
   isNewMemo: boolean;
@@ -54,41 +52,3 @@ function Memo(props: MemoProps) {
 }
 
 export default Memo;
-
-const StMemo = styled.div`
-  position: relative;
-
-  margin-right: 0.8rem;
-  padding: 2.8rem 3.2rem 2.8rem 3.2rem;
-  width: 67rem;
-
-  background-color: ${COLOR.SUB_BLUE_8};
-  border-radius: 2.5rem;
-
-  &:not(:hover) {
-    .dot {
-      opacity: 0;
-    }
-  }
-`;
-
-const StKeyword = styled.h1`
-  margin-bottom: 0.8rem;
-
-  color: ${COLOR.BLACK};
-  ${FONT_STYLES.SB_25_MEMO};
-`;
-
-const StContent = styled.p`
-  color: ${COLOR.GRAY_80};
-  ${FONT_STYLES.R_23_MEMO};
-
-  & > button {
-    color: ${COLOR.GRAY_30};
-    ${FONT_STYLES.R_23_MEMO};
-
-    &:hover {
-      color: ${COLOR.MAIN_BLUE};
-    }
-  }
-`;
