+++ src/components/learnDetail/memo/Memo/style.ts
@@ -1,0 +1,41 @@
+import styled from 'styled-components';
+import { COLOR } from '@src/styles/color';
+import { FONT_STYLES } from '@src/styles/fontStyle';
+
+export const StMemo = styled.div`
+  position: relative;
+
+  margin-right: 0.8rem;
+  padding: 2.8rem 3.2rem 2.8rem 3.2rem;
+  width: 67rem;
+
+  background-color: ${COLOR.SUB_BLUE_8};
+  border-radius: 2.5rem;
+
+  &:not(:hover) {
+    .dot {
+      opacity: 0;
+    }
+  }
+`;
+
+export const StKeyword = styled.h1`
+  margin-bottom: 0.8rem;
+
+  color: ${COLOR.BLACK};
+  ${FONT_STYLES.SB_25_MEMO};
+`;
+
+export const StContent = styled.p`
+  color: ${COLOR.GRAY_80};
+  ${FONT_STYLES.R_23_MEMO};
+
+  & > button {
+    color: ${COLOR.GRAY_30};
+    ${FONT_STYLES.R_23_MEMO};
+
+    &:hover {
+      color: ${COLOR.MAIN_BLUE};
+    }
+  }
+`;
