+++ src/components/learnDetail/memo/MemoForm/style.ts
@@ -1,0 +1,31 @@
+import { COLOR } from '@src/styles/color';
+import { FONT_STYLES } from '@src/styles/fontStyle';
+import styled from 'styled-components';
+
+export const StForm = styled.textarea`
+  padding: 0.8rem 0.8rem 0.8rem 1.2rem;
+  width: 60.6rem;
+
+  border: 2px solid ${COLOR.SUB_BLUE_50};
+  border-radius: 1.2rem;
+  background-color: transparent;
+
+  font-family: 'Pretendard';
+  ${FONT_STYLES.R_23_MEMO};
+  color: ${COLOR.GRAY_80};
+
+  overflow-y: hidden;
+  resize: none;
+  &:focus {
+    outline: none;
+  }
+`;
+
+export const StButtonContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+
+  width: 60.6rem;
+  height: 3rem;
+  margin-top: 1.2rem;
+`;
