+++ src/components/learnDetail/memo/MemoList/index.tsx
@@ -1,8 +1,7 @@
 import { HighlightData } from '@src/services/api/types/learn-detail';
-import { COLOR } from '@src/styles/color';
 import { useEffect, useState } from 'react';
-import styled from 'styled-components';
-import Memo from './Memo';
+import Memo from '../Memo';
+import { StMemoList } from './style';
 
 interface MemoListProps {
   highlightList: HighlightData[];
@@ -43,22 +42,3 @@ function MemoList(props: MemoListProps) {
 }
 
 export default MemoList;
-
-const StMemoList = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-
-  height: 35.8rem;
-  overflow-x: hidden;
-  overflow-y: auto;
-
-  &::-webkit-scrollbar {
-    width: 1rem;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background-color: ${COLOR.GRAY_10};
-    border-radius: 1.3rem;
-  }
-`;
