+++ src/components/learnDetail/memo/MemoList/style.ts
@@ -1,0 +1,21 @@
+import { COLOR } from '@src/styles/color';
+import styled from 'styled-components';
+
+export const StMemoList = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1.2rem;
+
+  height: 35.8rem;
+  overflow-x: hidden;
+  overflow-y: auto;
+
+  &::-webkit-scrollbar {
+    width: 1rem;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    background-color: ${COLOR.GRAY_10};
+    border-radius: 1.3rem;
+  }
+`;
