+++ src/components/learnDetail/memo/MemoPopupButton/index.tsx
@@ -1,8 +1,9 @@
 import ImageDiv from '@src/components/common/ImageDiv';
 import { icDotDefault, icDotHover } from 'public/assets/icons';
 import { useState, useRef, useEffect } from 'react';
-import styled from 'styled-components';
-import MemoPopup from './MemoPopup';
+
+import MemoPopup from '../MemoPopup';
+import { StMemoDotButton, StMemoDotImage } from './style';
 
 interface MemoPopupButtonProps {
   edit: () => void;
@@ -35,23 +36,3 @@ function MemoPopupButton(props: MemoPopupButtonProps) {
 }
 
 export default MemoPopupButton;
-
-const StMemoDotButton = styled.button`
-  position: absolute;
-  top: 1.2rem;
-  right: 1.2rem;
-  display: inline-block;
-
-  width: 4rem;
-  height: 4rem;
-  padding: 0;
-`;
-
-const StMemoDotImage = styled.div`
-  position: relative;
-
-  .dot {
-    position: absolute;
-    top: -2rem;
-  }
-`;
