+++ src/components/learnDetail/memo/MemoPopupButton/style.ts
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const StMemoDotButton = styled.button`
+  position: absolute;
+  top: 1.2rem;
+  right: 1.2rem;
+  display: inline-block;
+
+  width: 4rem;
+  height: 4rem;
+  padding: 0;
+`;
+
+export const StMemoDotImage = styled.div`
+  position: relative;
+
+  .dot {
+    position: absolute;
+    top: -2rem;
+  }
+`;
